Fix OsFileSystem::remove() so it works correctly when given a non-existent path. Simplify code in OsDirBase::getFileInfo(). Improve documentation of OsPathBase.
git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@17026 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXportLib/src/os/OsFileSystem.cpp
@@ -107,26 +107,28 @@ OsStatus OsFileSystem::remove(const OsPath& path, UtlBoolean bRecursive, UtlBool
     OsStatus retval = OS_INVALID;
     OsFileInfo info;
     OsPath testpath = path;
-    getFileInfo(testpath,info);
 
-    if (info.isDir())
+    if (getFileInfo(testpath, info) == OS_SUCCESS)
     {
-        if (bRecursive)
-        {
-            retval = removeTree(path,bForce);
-        }
-        else
-        {
-            if (rmdir((char *)path.data()) != -1)
+       if (info.isDir())
+       {
+          if (bRecursive)
+          {
+             retval = removeTree(path,bForce);
+          }
+          else
+          {
+             if (rmdir((char *)path.data()) != -1)
                 retval = OS_SUCCESS;
-        }
-    }
-    else
-    {
-       if (bForce)
-           setReadOnly(path,FALSE);
-       if (::remove(path.data()) != -1)
-           retval = OS_SUCCESS;
+          }
+       }
+       else
+       {
+          if (bForce)
+             setReadOnly(path,FALSE);
+          if (::remove(path.data()) != -1)
+             retval = OS_SUCCESS;
+       }
     }
 
     return retval;
